Cmap 6 support removed.

+++ Driver/Font/TrueType/FreeType/ttfile.h
@@ -149,7 +149,7 @@
   EXPORT_DEF
   TT_Error  TT_Read_At_File( STREAM_ARGS Long   position,
                                          void*  buffer,
-                                         Long   count );
+                                         Short  count );
 
   /* Return current file position */
